feat: form builder add model to document on specific profile

+++ components/com_emundus/models/settings.php
@@ -144,19 +144,25 @@ class EmundusModelsettings extends JModelList {
      * @since 1.0
      */
     function deleteTag($id) {
-        $db = $this->getDbo();
-        $query = $db->getQuery(true);
+		$deleted = false;
 
-        $query->delete($db->quoteName('#__emundus_setup_action_tag'))
-            ->where($db->quoteName('id') . ' = ' . $id);
+		if (!empty($id)) {
+			$db = $this->getDbo();
+			$query = $db->getQuery(true);
 
-        try {
-            $db->setQuery($query);
-            return $db->execute();
-        } catch(Exception $e) {
-            JLog::add('component/com_emundus/models/settings | Cannot delete the tag ' . $id . ' : ' . preg_replace("/[\r\n]/"," ",$query->__toString().' -> '.$e->getMessage()), JLog::ERROR, 'com_emundus');
-            return false;
-        }
+			$query->delete($db->quoteName('#__emundus_setup_action_tag'))
+				->where($db->quoteName('id') . ' = ' . $id);
+
+			try {
+				$db->setQuery($query);
+				$deleted=  $db->execute();
+			} catch(Exception $e) {
+				JLog::add('component/com_emundus/models/settings | Cannot delete the tag ' . $id . ' : ' . preg_replace("/[\r\n]/"," ",$query->__toString().' -> '.$e->getMessage()), JLog::ERROR, 'com_emundus');
+				$deleted = false;
+			}
+		}
+
+		return $deleted;
     }
 
     /**
